requirements.txt
# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0

# LLM and embeddings
langchain>=0.1.0
langchain-openai>=0.0.5  # For embeddings (required) + optional chat
openai>=1.0.0
langchain-anthropic>=0.1.0  # Optional: for Claude chat (install if using)

# Vector database
chromadb>=0.4.0  # Local development
pinecone>=8.0.0  # Optional: cloud vector DB (install if using)

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Rate limiting for scraping
ratelimit>=2.2.1

# Cloud storage (R2/S3)
boto3>=1.34.0
